+++ robotidy/api.py
@@ -7,7 +7,7 @@ from robotidy.app import Robotidy
 from robotidy.cli import TransformType, find_and_read_config, validate_regex
 from robotidy.disablers import RegisterDisablers
 from robotidy.files import DEFAULT_EXCLUDES
-from robotidy.utils import GlobalFormattingConfig
+from robotidy.utils import GlobalFormattingConfig, ROBOT_VERSION
 
 
 class RobotidyAPI(Robotidy):
@@ -42,6 +42,7 @@ class RobotidyAPI(Robotidy):
             check=False,
             output=output,
             force_order=False,
+            target_version=ROBOT_VERSION.major,
             color=False,
         )